reorganize uom
